Move result into ai folder
