+++ apps/desktop/src/lib/ai/result.ts
